+++ src/tox_poetry_dev_dependencies/__init__.py
@@ -2,8 +2,4 @@
 
 """tox-poetry-dev-dependencies."""
 
-from . import _meta
-
-__version__ = _meta.VERSION  # PEP 396
-
 # EOF