+++ test/test_unit.py
@@ -2,18 +2,18 @@
 
 """Unit tests."""
 
+import types
 import unittest
 
 import tox_poetry_dev_dependencies
 
 
-class TestProjectVersion(unittest.TestCase):
-    """Project version string."""
+class TestDummy(unittest.TestCase):
+    """Dummy."""
 
-    def test_project_has_version_string(self) -> None:
-        """Project should have a vesion string."""
-        self.assertIn('__version__', dir(tox_poetry_dev_dependencies))
-        self.assertIsInstance(tox_poetry_dev_dependencies.__version__, str)
+    def test_dummy(self) -> None:
+        """Dummy test."""
+        self.assertIsInstance(tox_poetry_dev_dependencies, types.ModuleType)
 
 
 # EOF